move schemas out of AddOrganisationFacility component, use one component for adding a facility/org
