+++ app/containers/Facilities/FacilitiesListContainer.tsx
@@ -2,7 +2,7 @@ import React, {useEffect, useState} from 'react';
 import {graphql, createRefetchContainer, RelayRefetchProp} from 'react-relay';
 import SearchTableLayout from 'components/SearchTableLayout';
 import createFacilityMutation from 'mutations/facility/createFacilityMutation';
-import AddFacility from 'components/facility/AddFacility';
+import AddOrganisationFacility from 'components/AddOrganisationFacility';
 import {FacilitiesListContainer_query} from 'FacilitiesListContainer_query.graphql';
 import FacilitiesRowItemContainer from './FacilitiesRowItemContainer';
 
@@ -74,7 +74,7 @@ export const FacilitiesList: React.FunctionComponent<Props> = props => {
         displayNameToColumnNameMap={displayNameToColumnNameMap}
         handleEvent={handleEvent}
       />
-      <AddFacility
+      <AddOrganisationFacility
         organisationRowId={organisation.rowId}
         reportingYear={getReportingYear.reportingYear}
         onAddFacility={handleAddFacility}
