+++ app/pages/user-dashboard.tsx
@@ -58,6 +58,7 @@ export default class UserDashBoard extends Component<Props> {
             // Relay requires a value here because it is set to not null, but the userId value is set via a trigger on insert
             userId: 1,
             organisationId: this.state.selectedOrg,
+            // TODO: set status on backend
             status: active ? 'APPROVED' : 'PENDING'
           }
         }
